requirements.txt
torch==2.0.1
torchvision==0.15.2
diffusers
accelerate
transformers
scipy
opencv-python
omegaconf
einops
git+https://github.com/pesser/splatting
roma
jaxtyping
boto3>=1.24.0
pytest>=7.0.0
Pillow>=9.0.0
boto3>=1.24.0
numpy>=1.23.0
pytest>=7.0.0
